+++ src/main/java/botLogic/Data.java
@@ -1,0 +1,9 @@
+package botLogic;
+import java.util.List;
+
+interface Data {
+    List<String> getSchedule(String id, int day);
+    void setSchedule(String group, List<List<String>> schedule);
+    void addUser(String id, String group);
+    void updateUser(String id, String group);
+}
